New tests to show issues with zero length dimensions in a multidimensional array.
I then patched HeapUtil to handle these. I am not sure if the actual byte
counts are correct regarding heap usage of such arrays, but in general these
will probably be of least concern to someone using HeapAudit. Would be nice
to eventually validate the numbers are correct for zero length arrays.